+++ lib/src/okta-angular.ts
@@ -10,11 +10,13 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-export { OktaAuthModule } from './okta/okta.module';
-export { OktaAuthGuard } from './okta/okta.guard';
-export { OktaConfig, OKTA_CONFIG, OKTA_AUTH } from './okta/models/okta.config';
+export { provideOktaAuth, withOktaConfig } from './okta/provide-okta';
+export type { OktaAuthFeatures } from './okta/provide-okta';
+export { canActivateAuthGuard, canActivateChildAuthGuard, canMatchAuthGuard } from './okta/okta.guard';
+export { OKTA_CONFIG, OKTA_AUTH } from './okta/models/okta.config';
+export type { OktaConfig } from './okta/models/okta.config';
 export { OktaAuthStateService } from './okta/services/auth-state.service';
-export { OktaAuthConfigService } from './okta/services/auth-config.serice';
+export { OktaAuthConfigService } from './okta/services/auth-config.service';
 export { OktaHasAnyGroupDirective } from './okta/has-any-group.directive';
 
 // Okta View Components