+++ test/spec/callback.component.test.ts
@@ -22,7 +22,6 @@ describe('OktaCallbackComponent', () => {
 
   beforeEach( () => {
     TestBed.configureTestingModule({
-      declarations: [OktaCallbackComponent],
       providers: [
         { provide: OktaAuthConfigService, useValue: configService },
         { provide: OKTA_AUTH, useValue: oktaAuth }
@@ -63,8 +62,7 @@ describe('OktaCallbackComponent', () => {
       fixture.detectChanges();
       await fixture.whenStable();
       expect(oktaAuth.idx.isInteractionRequiredError).toHaveBeenCalledWith(error);
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      expect(onAuthResume).toHaveBeenCalledWith(oktaAuth, (component as any).injector);
+      expect(onAuthResume).toHaveBeenCalledWith(oktaAuth, fixture.debugElement.injector);
       expect(component.error).toBe(undefined);
     });
 
@@ -77,8 +75,7 @@ describe('OktaCallbackComponent', () => {
       fixture.detectChanges();
       await fixture.whenStable();
       expect(oktaAuth.idx.isInteractionRequiredError).toHaveBeenCalledWith(error);
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      expect(onAuthRequired).toHaveBeenCalledWith(oktaAuth, (component as any).injector);
+      expect(onAuthRequired).toHaveBeenCalledWith(oktaAuth, fixture.debugElement.injector);
       expect(component.error).toBe(undefined);
     });